fix bench_dbms build
