+++ dbms/src/Functions/tests/bench_function_least.cpp
@@ -285,7 +285,7 @@ try
     auto context = DB::tests::TiFlashTestEnv::getContext();
     for (auto _ : state)
     {
-        executeFunction(context, func_name, col1, col2, col3);
+        executeFunction(*context, func_name, col1, col2, col3);
     }
 }
 CATCH
@@ -299,7 +299,7 @@ try
     auto context = DB::tests::TiFlashTestEnv::getContext();
     for (auto _ : state)
     {
-        executeFunction(context, func_name, col1, col2);
+        executeFunction(*context, func_name, col1, col2);
     }
 }
 CATCH
@@ -313,7 +313,7 @@ try
     auto context = DB::tests::TiFlashTestEnv::getContext();
     for (auto _ : state)
     {
-        executeFunction(context, func_name, col_nullable1, col_nullable2, col_nullable3);
+        executeFunction(*context, func_name, col_nullable1, col_nullable2, col_nullable3);
     }
 }
 CATCH
@@ -327,7 +327,7 @@ try
     auto context = DB::tests::TiFlashTestEnv::getContext();
     for (auto _ : state)
     {
-        executeFunction(context, func_name, col1, col2, col3);
+        executeFunction(*context, func_name, col1, col2, col3);
     }
 }
 CATCH
@@ -341,7 +341,7 @@ try
     auto context = DB::tests::TiFlashTestEnv::getContext();
     for (auto _ : state)
     {
-        executeFunction(context, func_name, col1, col2);
+        executeFunction(*context, func_name, col1, col2);
     }
 }
 CATCH
@@ -355,7 +355,7 @@ try
     auto context = DB::tests::TiFlashTestEnv::getContext();
     for (auto _ : state)
     {
-        executeFunction(context, func_name, col_nullable1, col_nullable2, col_nullable3);
+        executeFunction(*context, func_name, col_nullable1, col_nullable2, col_nullable3);
     }
 }
 CATCH
@@ -370,7 +370,7 @@ try
     auto context = DB::tests::TiFlashTestEnv::getContext();
     for (auto _ : state)
     {
-        executeFunction(context, func_name, v_col);
+        executeFunction(*context, func_name, v_col);
     }
 }
 CATCH
@@ -384,7 +384,7 @@ try
     auto context = DB::tests::TiFlashTestEnv::getContext();
     for (auto _ : state)
     {
-        executeFunction(context, func_name, v_col);
+        executeFunction(*context, func_name, v_col);
     }
 }
 CATCH
